chore(core): refactoring - handler is moved from lambda to be reused for means of deployment other than aws lambda.
