+++ core/cmd/lambda/main.go
@@ -1,13 +1,17 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
 	"log"
+	"net/http"
 	"os"
 	"strconv"
 
+	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/kislerdm/diagramastext/core"
+	coreHandler "github.com/kislerdm/diagramastext/core/handler"
 )
 
 func main() {
@@ -40,4 +44,80 @@ func mustParseInt(s string) int {
 func mustParseFloat32(s string) float32 {
 	o, _ := strconv.ParseFloat(s, 10)
 	return float32(o)
+}
+
+type corsHeaders map[string]string
+
+func (h corsHeaders) setHeaders(resp events.APIGatewayProxyResponse) events.APIGatewayProxyResponse {
+	if h == nil {
+		return resp
+	}
+
+	if resp.Headers == nil {
+		resp.Headers = map[string]string{}
+	}
+
+	for k, v := range h {
+		resp.Headers[k] = v
+
+		if k == "Access-Control-Allow-Origin" && (v == "" || v == "'*'") {
+			resp.Headers[k] = "*"
+		}
+	}
+
+	return resp
+}
+
+func handler(
+	clientModel core.ClientInputToGraph, clientDiagram core.ClientGraphToDiagram, corsHeaders corsHeaders,
+) func(
+	ctx context.Context, req events.APIGatewayProxyRequest,
+) (events.APIGatewayProxyResponse, error) {
+	return func(
+		ctx context.Context, req events.APIGatewayProxyRequest,
+	) (events.APIGatewayProxyResponse, error) {
+		prompt, err := coreHandler.ReadPrompt([]byte(req.Body))
+		if err != nil {
+			return corsHeaders.setHeaders(
+				events.APIGatewayProxyResponse{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       "could not recognise the prompt format",
+				},
+			), err
+		}
+
+		if err := coreHandler.ValidatePrompt(prompt); err != nil {
+			return corsHeaders.setHeaders(
+				events.APIGatewayProxyResponse{
+					StatusCode: http.StatusBadRequest,
+					Body:       err.Error(),
+				},
+			), err
+		}
+
+		graph, err := clientModel.Do(ctx, prompt)
+		if err != nil {
+			return corsHeaders.setHeaders(parseClientError(err)), err
+		}
+
+		svg, err := clientDiagram.Do(ctx, graph)
+		if err != nil {
+			return corsHeaders.setHeaders(parseClientError(err)), err
+		}
+
+		return corsHeaders.setHeaders(
+			events.APIGatewayProxyResponse{
+				StatusCode: http.StatusOK,
+				Body:       string(svg.MustMarshal()),
+			},
+		), err
+	}
+}
+
+func parseClientError(err error) events.APIGatewayProxyResponse {
+	v := coreHandler.ParseClientError(err)
+	return events.APIGatewayProxyResponse{
+		StatusCode: v.StatusCode,
+		Body:       string(v.Body),
+	}
 }
